acabado el crud de llit y poner confirmación al eliminar pacientes

+++ VitalTech/src/app/models/pacientes/pacientes.component.ts
@@ -50,13 +50,15 @@ export class PacientesComponent {
 
   deletePacient(id: string): void {
 
-    this.pacienteService.deletePacient(id).subscribe({
-      error: error => alert('ERROR, el pacient encara té episodis médics'),
-      complete: () => {
-        alert('Pacient Borrat'),
+    if(confirm('Esta seguro de eliminar este paciente?')) {
+      this.pacienteService.deletePacient(id).subscribe({
+        error: error => alert('ERROR, el pacient encara té episodis médics'),
+        complete: () => {
+          alert('Pacient Borrat'),
           this.loadPacients()
-      }
-    });
+        }
+      });
+    }
   }
 
   updatePacient(idPacient: string) {
